Added Computer class and update to Memory class
The update to memory class was more of a finish from last nights work. I
added a new class (Computer) that I think will help keep everything
together. My vision for this class is that this will be the class that
brings in everything and will easily control the finish system.